Updates to scanning logic/entrypoints for gcobra
This commit intends to massively simplify entrypoints to gcobra.

The idea is that in many cases, people "just want some flags and
positionals", because their program IS their command. Therefore,
they should be able to scan only a struct of positionals/options,
and be given a cobra.Command in return, which they can directly
use as an entrypoint with cmd.Execute()

New entrypoint `Parse(interface{}) *cobra.Command` is now accepting
any interface, so that it can either scan a group of options &
positionals not stored into a Commander, and will return a command
that is comprising all of these.
But if the client passes a struct implementing Commander, the command
returned will be wrapped around the struct' implementation, as usual.

All other features are kept, and recursive scanning will still work.
A few things should "fixed" or decided upon, some defaults for instance.