Modifications and fixes in gen/gcobra scan code
There was a problem in that all the scan handlers would not correctly
work in sync with the original sflags functions. The result was, for
instance, commands getting scanned twice, because the parent scanner
would keep scanning nested structs even if it deemed them to be a
subcommand (thus should pass the work to a new scanner).
This is fixed because:

- The internal/scan code has been fixed here and there: sometimes it
  would badly signal what were its actions.
- the gen/gcobra/group.go file has now a small scan handler dedicated
  to scanning a field for flags BY DEFAULT. So for each field, now:
  We test for a positional struct, a command struct, a group struct,
  and if these three tests have failed, we finally scan the field as
  a single flag.